backend_service/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.5.0
celery[redis]
pydantic>=2.0.0
pydantic-settings
python-dotenv